feat: refactor DTO imports to use auth package for improved organization
